Curso/src/curso/g10/junit0/Debito.java
package curso.g10.junit0;

import java.util.Calendar;

public class Debito extends Tarjeta {
	public Debito(String numero, String titular, Calendar fecha) {
		super(numero, titular, fecha);
	}

	public void retirar(double x) throws Exception {
		this.mCuentaAsociada.retirar("Retirada en cajero automático", x);
	}

	public void ingresar(double x) throws Exception {
		this.mCuentaAsociada.retirar("Ingreso en cajero automático", x);
	}

	public void pagoEnEstablecimiento(String datos, double x) throws Exception {
		this.mCuentaAsociada.retirar("Compra en :" + datos, x);
	}

	public double getSaldo() {
		return mCuentaAsociada.getSaldo();
	}
}
